Trying to send more precise location data.

+++ src/ingameelements/movingentity.cpp
@@ -32,16 +32,16 @@ void MovingEntity::interpolate(Entity *prev_entity, Entity *next_entity, double 
 
 void MovingEntity::serialize(Gamestate *state, WriteBuffer *buffer, bool fullupdate)
 {
-    buffer->write<float>(x);
-    buffer->write<float>(y);
+    buffer->write<double>(x);
+    buffer->write<double>(y);
     buffer->write<float>(hspeed);
     buffer->write<float>(vspeed);
 }
 
 void MovingEntity::deserialize(Gamestate *state, ReadBuffer *buffer, bool fullupdate)
 {
-    x = buffer->read<float>();
-    y = buffer->read<float>();
+    x = buffer->read<double>();
+    y = buffer->read<double>();
     hspeed = buffer->read<float>();
     vspeed = buffer->read<float>();
 }
